Moved captcha.py into the helpers folder
